tracing: Make duplicate count from tracing_map available
Though extremely rare, there can be duplicate entries in the tracing
map.  This isn't normally a problem, as the sorting code makes this
transparent by merging them during the sort.

It's useful to know however, as a check on that assumption - if a
non-zero duplicate count is seen more than rarely, it might indicate
an unexpected change to the algorithm, or a pathological data set.

Add an extra param to tracing_map_sort_entries() and use it to display
the value in the hist trigger output.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>