tracing: Introduce tags for histogram triggers
Introduce the 'tag=mygrouping' histogram specifier, allowing groups of
related hist triggers to be defined.  This is useful for managing
groups of hist triggers and making it easier to apply operations to
sets of triggers matching a specified tag name.  Examples would be
removing a set of triggers as a group, or deleting matching records
for every trigger with the same tag, using the delete() operation.

Currently, there can only be one tag associated with any given
histogram trigger.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>