tracing: Implement delete for tracing_map.
Some of the use cases of hist triggers, for example tracing network
packets by their address, will mean that the tracing_map will inevitably
fill up. This patch implement delete to be able to delete and free up
entries for future elements.

There are few new flags added. There are deleted flags added
to tracing_map_elt and tracing_map_entry structs. Both of them save the
same purpose -- to indicate whether the entry/elt has been deleted.
There are other flags/variable to keep track of deleted elts. These are
has_freed_elt, last_freed_elt, find_failed.

When it is invoked, tracing_map_delete() just sets entry->deleted and
sets entry->val to NULL.  The element is cleared out later by the RCU
callback (free_elt()). The callback will also set the deleted flag for
the elt. This indicates the element is ready to be used again.

When a new element is inserted it may be inserted in an entry which has
been deleted before (entry->deleted is set). If so, it has to make sure
that there is no element with the same key by traversing the array until
an empty non-deleted (entry->deleted is clear) element is found. When it
is looking for a new elt (using get_free_elt()), it will first try to
use elts in order.  Only if elements have exhausted, it will go looking
for elements with deleted flag set.

The RCU is really protecting the entry->elt pointer, and not the
contents which are stored in the elt. the tracing_map_insert() and
tracing_map_delete() are the RCU updaters and tracing_map_lookup() will
be the RCU reader.

Credits-to: Tom Zanussi <tom.zanussi@linux.intel.com>
Signed-off-by: Vedang Patel <vedang.patel@intel.com>