tracing: Implement action support for delete.
This patch exposes the shiney new delete functionality added to the
tracing_map hash table to the world. It does so via the delete() action.

The only requirement to delete from a hist trigger is that it should
have a tag assigned (using "tag=" while defining the hist trigger).
Following is a simple example to use the API:

echo 'hist:tag=trig1:keys=pid:ts0=$common_timestamp.usecs' >
/sys/kernel/debug/tracing/events/sched/sched_wakeup/trigger
echo 'hist:keys=pid:size=131072:delete(trig1)' >
/sys/kernel/debug/tracing/events/sched/sched_process_exit/trigger

The hist trigger on which the delete is being executed (in the above
example, trig1) cannot be deleted before the one with the delete action
is deleted.

Signed-off-by: Vedang Patel <vedang.patel@intel.com>
Co-Developed-by: Tom Zanussi <tom.zanussi@linux.intel.com>
Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>