tracing: Add resolve_var_refs bitmap
For variables that are both defined and used on the same event, the
hist entry the variables are attached to doesn't exist yet, so
resolve_var_refs() obviously won't be able to read them.  Since
resolve_var_refs is really just a cache for values read from the
variables, in the case we don't have an entry in the cache for a given
variable reference, we can just read it directly when retrieving the
variable reference rather than use the bogus value in the array.  In
order to know that a value in the array wasn't resolved, we add a
bitmap to track each entry.  Because a resolved value will only ever
be missing if the case of a variable defined on the same event we've
now added and are updating, we don't need to specify a hist_data as
well since that's always 'this'.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>