tracing: Refactor action code
The action code is really of the form handler.action, where both
handler and action can be combined in arbitrary ways e.g. in addition
to the current onmatch().tracexxx() and onmax().snapshot, we should be
able to do onmax().tracexxx() or onmatch().snapshot(), etc.

Refactor it all to allow arbitrary combinations, which also happens to
regularize and clean up the code in the process.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>